App.jsx refactoring
