--- src/context/DataContext.js
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const DataContext = createContext(null);
